move collect_sourcemap_artifacts
