+++ crates/symbolicator-service/src/services/symbolication/sourcemap.rs
@@ -1,12 +1,8 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 
-use reqwest::Url;
-use sourcemap::locate_sourcemap_reference;
 use symbolic::sourcemapcache::{ScopeLookupResult, SourcePosition};
-use tempfile::NamedTempFile;
-use url::Position;
 
-use crate::services::{download::sentry::SearchArtifactResult, sourcemap::OwnedSourceMapCache};
+use crate::services::sourcemap::OwnedSourceMapCache;
 use crate::types::{
     CompleteJsStacktrace, FrameStatus, JsProcessingCompletedSymbolicationResponse,
     JsProcessingRawFrame, JsProcessingRawStacktrace, JsProcessingSymbolicatedFrame,
@@ -19,67 +15,12 @@ use super::{JsProcessingSymbolicateStacktraces, SymbolicationActor};
 
 // TODO(sourcemap): Move this and related functions to its own file, similar to how `ModuleLookup` works.
 impl SymbolicationActor {
-    // TODO(sourcemap): Handle 3rd party servers fetching (payload should decide about it, as it's user-configurable in the UI)
-    async fn collect_stacktrace_artifacts(
-        &self,
-        request: &JsProcessingSymbolicateStacktraces,
-    ) -> HashMap<String, OwnedSourceMapCache> {
-        let mut unique_abs_paths = HashSet::new();
-        for stacktrace in &request.stacktraces {
-            for frame in &stacktrace.frames {
-                unique_abs_paths.insert(frame.abs_path.clone());
-            }
-        }
-
-        let release_archive = self.sourcemaps.list_artifacts(request.source.clone()).await;
-
-        let compute_caches = unique_abs_paths.into_iter().map(|abs_path| async {
-            let abs_path_url = Url::parse(&abs_path).ok()?;
-
-            let source_artifact = get_release_file_candidate_urls(&abs_path_url)
-                .into_iter()
-                .find_map(|candidate| release_archive.get(&candidate))?;
-            let source_file = self
-                .sourcemaps
-                .fetch_artifact(request.source.clone(), source_artifact.id.clone())
-                .await?;
-
-            // TODO(sourcemap): Report missing sourcemap url error
-            let sourcemap_url =
-                resolve_sourcemap_url(&abs_path_url, source_artifact, &source_file)?;
-
-            let sourcemap_artifact = get_release_file_candidate_urls(&sourcemap_url)
-                .into_iter()
-                .find_map(|candidate| release_archive.get(&candidate))?;
-            // TODO(sourcemap): Report missing source error
-            let sourcemap_file = self
-                .sourcemaps
-                .fetch_artifact(request.source.clone(), sourcemap_artifact.id.clone())
-                .await?;
-
-            let cache = self
-                .sourcemaps
-                .fetch_cache(&sourcemap_file, &sourcemap_file)
-                .await
-                // TODO: properly report errors here
-                .unwrap();
-
-            Some((abs_path, cache))
-        });
-
-        futures::future::join_all(compute_caches)
-            .await
-            .into_iter()
-            .flatten()
-            .collect()
-    }
-
     #[tracing::instrument(skip_all)]
     pub async fn js_processing_symbolicate(
         &self,
         request: JsProcessingSymbolicateStacktraces,
     ) -> Result<JsProcessingCompletedSymbolicationResponse, anyhow::Error> {
-        let artifacts = self.collect_stacktrace_artifacts(&request).await;
+        let artifacts = self.sourcemaps.collect_stacktrace_artifacts(&request).await;
 
         let stacktraces: Vec<_> = request
             .stacktraces
@@ -174,55 +115,6 @@ fn js_processing_symbolicate_frame(
     }
 }
 
-// Transforms a full absolute url into 2 or 4 generalized options. Based on `ReleaseFile.normalize`.
-// https://github.com/getsentry/sentry/blob/master/src/sentry/models/releasefile.py
-fn get_release_file_candidate_urls(url: &Url) -> Vec<String> {
-    let mut urls = vec![];
-
-    // Absolute without fragment
-    urls.push(url[..Position::AfterQuery].to_string());
-
-    // Absolute without query
-    if url.query().is_some() {
-        urls.push(url[..Position::AfterPath].to_string())
-    }
-
-    // Relative without fragment
-    urls.push(format!(
-        "~{}",
-        &url[Position::BeforePath..Position::AfterQuery]
-    ));
-
-    // Relative without query
-    if url.query().is_some() {
-        urls.push(format!(
-            "~{}",
-            &url[Position::BeforePath..Position::AfterPath]
-        ));
-    }
-
-    urls
-}
-
-// Joins together frames `abs_path` and discovered sourcemap reference.
-fn resolve_sourcemap_url(
-    abs_path: &Url,
-    source_artifact: &SearchArtifactResult,
-    mut source_file: &NamedTempFile,
-) -> Option<Url> {
-    if let Some(header) = source_artifact.headers.get("Sourcemap") {
-        abs_path.join(header).ok()
-    } else if let Some(header) = source_artifact.headers.get("X-SourceMap") {
-        abs_path.join(header).ok()
-    } else {
-        use std::io::Seek;
-
-        let sm_ref = locate_sourcemap_reference(source_file.as_file()).ok()??;
-        source_file.rewind().ok()?;
-        abs_path.join(sm_ref.get_url()).ok()
-    }
-}
-
 /// Fold multiple consecutive occurences of the same property name into a single group, excluding the last component.
 ///
 /// foo | foo
@@ -273,48 +165,6 @@ fn fold_function_name(function_name: &str) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_get_release_file_candidate_urls() {
-        let url = "https://example.com/assets/bundle.min.js".parse().unwrap();
-        let expected = vec![
-            "https://example.com/assets/bundle.min.js",
-            "~/assets/bundle.min.js",
-        ];
-        assert_eq!(get_release_file_candidate_urls(&url), expected);
-
-        let url = "https://example.com/assets/bundle.min.js?foo=1&bar=baz"
-            .parse()
-            .unwrap();
-        let expected = vec![
-            "https://example.com/assets/bundle.min.js?foo=1&bar=baz",
-            "https://example.com/assets/bundle.min.js",
-            "~/assets/bundle.min.js?foo=1&bar=baz",
-            "~/assets/bundle.min.js",
-        ];
-        assert_eq!(get_release_file_candidate_urls(&url), expected);
-
-        let url = "https://example.com/assets/bundle.min.js#wat"
-            .parse()
-            .unwrap();
-        let expected = vec![
-            "https://example.com/assets/bundle.min.js",
-            "~/assets/bundle.min.js",
-        ];
-        assert_eq!(get_release_file_candidate_urls(&url), expected);
-
-        let url = "https://example.com/assets/bundle.min.js?foo=1&bar=baz#wat"
-            .parse()
-            .unwrap();
-        let expected = vec![
-            "https://example.com/assets/bundle.min.js?foo=1&bar=baz",
-            "https://example.com/assets/bundle.min.js",
-            "~/assets/bundle.min.js?foo=1&bar=baz",
-            "~/assets/bundle.min.js",
-        ];
-        assert_eq!(get_release_file_candidate_urls(&url), expected);
-    }
-
     #[test]
     fn test_fold_function_name() {
         assert_eq!(fold_function_name("foo"), "foo");
